Complete bloglist 4.21

+++ part4/bloglist/backend/controllers/blogs.js
@@ -18,19 +18,22 @@ blogsRouter.get("/:id", async (req, res) => {
 });
 
 blogsRouter.post("/", async (req, res) => {
-  const body = req.body;
-  if (!body) {
-    throw new AppError("Missing content", 400);
+  if (!req.token) {
+    throw new AppError("missing credentials", 401);
+  }
+  const { title, author, url, likes } = req.body;
+  if (!title || !url) {
+    throw new AppError("missing content", 400);
   }
 
-  const userData = jwt.verify(req.token, process.env.JWT_SECRET_KEY);
-  if (!userData) {
-    throw new AppError("Invalid credentials", 401);
+  const decodedData = jwt.verify(req.token, process.env.JWT_SECRET_KEY);
+
+  const user = await User.findById(decodedData.id);
+  if (!user) {
+    throw new AppError("invalid credentials", 401);
   }
 
-  const user = await User.findById(userData.id);
-
-  const blog = new Blog({ ...body, user: user.id });
+  const blog = new Blog({ title, author, url, likes, user: user.id });
   const savedBlog = await blog.save();
 
   user.blogs = user.blogs.concat(savedBlog.id);
@@ -43,7 +46,7 @@ blogsRouter.put("/:id", async (req, res) => {
   const blogId = req.params.id;
   const blogToUpdate = await Blog.findById(blogId);
   if (!blogToUpdate) {
-    throw new AppError("Person not found", 404);
+    throw new AppError("person not found", 404);
   }
   blogToUpdate.likes = req.body.likes;
   const updatedBlog = await blogToUpdate.save();
@@ -51,8 +54,28 @@ blogsRouter.put("/:id", async (req, res) => {
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
+  if (!req.token) {
+    throw new AppError("missing credentials", 401);
+  }
+
   const blogId = req.params.id;
+  const blogToDelete = await Blog.findById(blogId);
+  if (!blogToDelete) {
+    throw new AppError("blog not found", 404);
+  }
+
+  const decodedData = jwt.verify(req.token, process.env.JWT_SECRET_KEY);
+
+  if (blogToDelete.user.toString() !== decodedData.id) {
+    throw new AppError("not authorized to delete this blog", 403);
+  }
+
   await Blog.findByIdAndDelete(blogId);
+
+  const user = await User.findById(decodedData.id);
+  user.blogs = user.blogs.filter((b) => b.toString() !== blogId);
+  await user.save();
+
   return res.status(204).end();
 });
